// src/types.ts
export interface Service {
    id: number;
    service?: number;
    category: string;
    name: string;
    type: ServiceType;
    rate: number;
    original_rate: number;
    min: number;
    max: number;
    averageTime: string;
    refill: boolean;
    cancel: boolean;
    platform_id?: SocialPlatform;
    custom_description?: string | null;
}
// ...
export type ServiceType =
    | 'Default'
    | 'Custom Comments'
    | 'Custom Comments Package'
    | 'Mentions with Hashtags'
    | 'Package'
    | 'Poll';

export type OrderStatus =
    | 'pending'
    | 'processing'
    | 'in_progress'
    | 'completed'
    | 'cancelled'
    | 'partial';

export interface CustomField {
    type: 'text' | 'link' | 'comment' | 'hashtag' | 'note';
    label?: string;
    value: string;
}
// ...
export type SocialPlatform =
    | 'instagram'
    | 'tiktok'
    | 'youtube'
    | 'facebook'
    | 'twitter'
    | 'telegram'
    | 'other'
    | 'top';